feat: refactoring

+++ src/main/java/com/onlyoffice/registry/service/BasicWorkspaceService.java
@@ -1,5 +1,6 @@
 package com.onlyoffice.registry.service;
 
+import com.onlyoffice.registry.InvalidRegistryOperationException;
 import com.onlyoffice.registry.dto.WorkspaceDTO;
 import com.onlyoffice.registry.mapper.WorkspaceMapper;
 import com.onlyoffice.registry.model.Workspace;
@@ -22,20 +23,23 @@ public class BasicWorkspaceService implements WorkspaceService {
 
     @Transactional(
             isolation = Isolation.READ_UNCOMMITTED,
-            rollbackFor = {TransactionException.class},
-            timeout = 2
+            readOnly = true
     )
     @Cacheable("workspaces")
     public WorkspaceDTO getWorkspace(WorkspaceID workspaceID) {
         log.debug("trying to get workspace of type: {} with id: {}", workspaceID.getWorkspaceType(), workspaceID.getWorkspaceId());
         return WorkspaceMapper.INSTANCE.toDTO(this.workspaceRepository
                 .findById(workspaceID)
-                .orElseThrow(() -> new RuntimeException("Could not get: Workspace with this id and type does not exist")));
+                .orElseThrow(() -> new InvalidRegistryOperationException("Could not get: Workspace with this id and type does not exist")));
     }
 
     @Transactional(
             isolation = Isolation.SERIALIZABLE,
-            rollbackFor = {TransactionException.class, RuntimeException.class},
+            rollbackFor = {
+                    TransactionException.class,
+                    InvalidRegistryOperationException.class,
+                    RuntimeException.class
+            },
             timeout = 2
     )
     public WorkspaceDTO saveWorkspace(String workspaceTypeName, WorkspaceDTO workspaceDTO) {
@@ -43,37 +47,40 @@ public class BasicWorkspaceService implements WorkspaceService {
         Workspace workspace = WorkspaceMapper.INSTANCE.toEntity(workspaceDTO);
         workspace.getId().setWorkspaceType(workspaceTypeName);
         if (this.workspaceRepository.existsById(workspace.getId()))
-            throw new RuntimeException("Could not save: Workspace with this id and workspace type already exists");
+            throw new InvalidRegistryOperationException("Could not save: Workspace with this id and workspace type already exists");
         this.workspaceRepository.save(workspace);
         return workspaceDTO;
     }
 
     @Transactional(
             isolation = Isolation.SERIALIZABLE,
-            rollbackFor = {TransactionException.class, RuntimeException.class},
+            rollbackFor = {
+                    TransactionException.class,
+                    InvalidRegistryOperationException.class,
+                    RuntimeException.class
+            },
             timeout = 3
     )
     @CacheEvict("workspaces")
     public void deleteWorkspace(WorkspaceID workspaceID) {
         log.debug("trying to delete workspace of type: {} with id: {}", workspaceID.getWorkspaceType(), workspaceID.getWorkspaceId());
-        try {
-            this.workspaceRepository.deleteById(workspaceID);
-        } catch (Exception e) {
-            throw new RuntimeException("Could not delete: Workspace with this id and workspace type does not exist");
-        }
+        if (!this.workspaceRepository.existsById(workspaceID))
+            throw new InvalidRegistryOperationException("Could not delete: Workspace with this id and workspace type does not exist");
+        this.workspaceRepository.deleteById(workspaceID);
     }
 
     @Transactional(
             isolation = Isolation.SERIALIZABLE,
-            rollbackFor = {TransactionException.class, RuntimeException.class},
+            rollbackFor = {
+                    TransactionException.class,
+                    InvalidRegistryOperationException.class,
+                    RuntimeException.class
+            },
             timeout = 3
     )
     @CacheEvict("workspaces")
     public void deleteAllWorkspacesByType(String workspaceTypeName) {
-        try {
-            this.workspaceRepository.deleteAllByIdWorkspaceType(workspaceTypeName);
-        } catch (RuntimeException e) {
-            throw new RuntimeException("Could not delete all workspaces of type "+workspaceTypeName);
-        }
+        log.debug("trying to remove all workspaces of type: {}", workspaceTypeName);
+        this.workspaceRepository.deleteAllByIdWorkspaceType(workspaceTypeName);
     }
 }
